requirements.txt
fastapi
uvicorn
python-dotenv
agno
python-telegram-bot
requests
